Add alternate method of setting up pipes
TypeScript seems to have this really jank commonjs export pattern, which
means that in order to use a class in JS one has to use multiple 'new's
and it just gets very messy.

This creates a "Plumber" interface, which is essentially a sort of Pipe
factory. It gets called, returns a new Pipe... Bypassing all of the
weird ass prototypal inheritance thingies!